test(yaml): add taint mode parsing tests and test helpers
- Introduced a new test file `taint_tests.rs` with unit tests for taint rule parsing.
- Moved shared parsing test helpers (`first_err_diagnostic` and `check_first_rule`) to a new module `test_helpers.rs`.
- Updated `parser_tests.rs` to use the new test helpers and remove taint-related tests.
- Updated test module declarations in `mod.rs` to include the new `taint_tests` and `test_helpers` modules.

These changes improve test organization and coverage for taint mode YAML parsing.
