+++ crates/sempai-yaml/src/tests/mod.rs
@@ -2,3 +2,5 @@
 
 mod behaviour;
 mod parser_tests;
+mod taint_tests;
+mod test_helpers;
